Hoshpital application
class Hospital is given to you. Implement the following features in the class.
Variables : 
private : hospitalCode int, hospitalName String,
 listOfTreatments List<String>,contactPerson String,
 contactNumber String,location String
public : Constructors:
Parameterized Constructor with arguments in the order of 
hospitalName,listOfTreatments,
 contactPerson,contactNumber,location
Methods : public
 Generate All Getters & Setters 
 Generate toString() to return a formatted string as per the below example.
Hospital [hospitalCode=101, hospitalName=YASHODA, listOfTreatments=[Cardiac, ENT, 
Ortho, Pediatric, Gastro], 
contactPerson=MATHEWS, contactNumber=9848222222, location=Sec]
 
A class HospitalService is given to you. Implement the following features in the class.
Methods : public 
addHospital(): This method should receive hospital details 
 as arguments and add to ArrayList of type Hospital and return the generated hospital 
code.
Note : Hospital Code is not given as argument but it should be 
 autogenerated while creating object. It should start from 100.
 getHospitals() : This method should return existing hospitals 
names and its codes in the form of Map variable that contains Hospital 
Code as key (Integer) and Hospital name as Value (String).
 getHospitalDetails() : This method should receive hospital code 
as argument and return the matching object of hospital after searching in 
the ArrayList. Return null if object not found.
A class HospitalFinder is given to you with a main method. 
Create multiple Hospital objects and test with the methods of HospitalService class here